Return values over buffers; int pathfinding costs
UnitsOnTeam/FindAdjacentOpponents/BuildTurn return List<T>. Pathfinder
costs are int. New GridDirections.Orthogonal replaces 3 duplicated
arrays. CombatMath takes UnitStats.

+++ Assets/Scripts/Utility/ITerrainCostSource.cs
@@ -1,9 +1,8 @@
 using UnityEngine;
 
-/// <summary>Read-only terrain query Pathfinder needs — implemented by GridManager,
-/// kept as a tiny interface so Pathfinder doesn't depend on a MonoBehaviour.</summary>
+/// <summary>Read-only terrain query Pathfinder needs, implemented by GridManager — kept as a tiny interface so Pathfinder doesn't depend on a MonoBehaviour.</summary>
 public interface ITerrainCostSource
 {
     bool Contains(Vector2Int cell);
-    float MovementCost(Vector2Int cell);
+    int MovementCost(Vector2Int cell);
 }
